python/requirements.txt
scikit-learn
numpy
matplotlib
pandas
mypy
flake8
black
pyfunctional
scikit-image
pyyaml
jupyter
scipy